# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation Api Gutter Runner",
        "description": "Implémentation de Swagger",
        "contact": {
            "email": "admin@app.gutterrunner.ca"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Serveur d'API de démonstration"
        }
    ],
    "paths": {
...
        "/api/demande_product_presentations": {
            "get": {
                "tags": [
                    "Demandes"
                ],
                "summary": "Liste des présentations de produits",
                "description": "Liste des présentations de produits",
                "operationId": "productPresentationsList",
                "responses": {
                    "200": {
                        "description": "Liste obtenue",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "ID",
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "description": "Nom",
                                                        "type": "string",
                                                        "example": "GutterClean System"
                                                    },
                                                    "url": {
                                                        "description": "Url",
                                                        "type": "string",
                                                        "example": "https://www.youtube.com/watch?v=Zvp7E5vMrWM"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "total": {
                                            "description": "Nombre total",
                                            "type": "integer",
                                            "example": 3
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Utilisateur non connecté",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "L'utilisateur n'est même pas connecté",
                                            "type": "string",
                                            "example": "Vous ne vous êtes pas encore authentifié."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/demande_sequences": {
            "get": {
                "tags": [
                    "Demandes"
                ],
                "summary": "Numéro de séquence suivant des demandes",
                "description": "Numéro de séquence suivant des demandes",
                "operationId": "demandeSequence",
                "responses": {
                    "200": {
                        "description": "Liste obtenue",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "next_sequence": {
                                                    "description": "Numéro de séquence suivant pour la création d'une demande",
                                                    "type": "string",
                                                    "example": "010000"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Utilisateur non connecté",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "L'utilisateur n'est même pas connecté",
                                            "type": "string",
                                            "example": "Vous ne vous êtes pas encore authentifié."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/demande_services": {
            "get": {
                "tags": [
                    "Demandes"
                ],
                "summary": "Liste des services de demande",
                "description": "Liste des services de demande",
                "operationId": "demandeServicesList",
                "responses": {
                    "200": {
                        "description": "Liste obtenue",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "ID",
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "description": "Nom",
                                                        "type": "string",
                                                        "example": "Installation"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "total": {
                                            "description": "Nombre total",
                                            "type": "integer",
                                            "example": 7
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Utilisateur non connecté",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "L'utilisateur n'est même pas connecté",
                                            "type": "string",
                                            "example": "Vous ne vous êtes pas encore authentifié."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}